+++ backend/dataall/modules/dataset_sharing/services/dataset_sharing_alarm_service.py
@@ -3,13 +3,13 @@ from datetime import datetime
 
 from dataall.core.environment.db.environment_models import Environment
 from dataall.modules.dataset_sharing.db.share_object_models import ShareObject
-from dataall.modules.datasets_base.db.dataset_models import DatasetTable, Dataset, DatasetStorageLocation, DatasetBucket
+from dataall.modules.datasets_base.db.dataset_models import DatasetTable, DatasetStorageLocation, DatasetBucket
 from dataall.base.utils.alarm_service import AlarmService
 
 log = logging.getLogger(__name__)
 
 
-class DatasetAlarmService(AlarmService):
+class DatasetSharingAlarmService(AlarmService):
     """Contains set of alarms for datasets"""
 
     def trigger_table_sharing_failure_alarm(
@@ -69,24 +69,6 @@ class DatasetAlarmService(AlarmService):
         - AWS Account:                {target_environment.AwsAccountId}
         - Region:                            {target_environment.region}
         - Glue Database:              {table.GlueDatabaseName}shared
-    """
-        return self.publish_message_to_alarms_topic(subject, message)
-
-    def trigger_dataset_sync_failure_alarm(self, dataset: Dataset, error: str):
-        log.info(f'Triggering dataset {dataset.name} tables sync failure alarm...')
-        subject = f'Data.all Dataset Tables Sync Failure for {dataset.name}'[:100]
-        message = f"""
-You are receiving this email because your Data.all {self.envname} environment in the {self.region} region has entered the ALARM state, because it failed to synchronize Dataset {dataset.name} tables from AWS Glue to the Search Catalog.
-
-Alarm Details:
-    - State Change:               	OK -> ALARM
-    - Reason for State Change:      {error}
-    - Timestamp:                              {datetime.now()}
-    Dataset
-     - Dataset URI:                   {dataset.datasetUri}
-     - AWS Account:                {dataset.AwsAccountId}
-     - Region:                            {dataset.region}
-     - Glue Database:              {dataset.GlueDatabaseName}
     """
         return self.publish_message_to_alarms_topic(subject, message)
