+++ backend/dataall/modules/dataset_sharing/services/share_managers/lf_share_manager.py
@@ -19,7 +19,7 @@ from dataall.modules.dataset_sharing.services.dataset_sharing_enums import (
     ShareItemHealthStatus,
 )
 from dataall.modules.datasets_base.db.dataset_models import DatasetTable, Dataset
-from dataall.modules.dataset_sharing.services.dataset_alarm_service import DatasetAlarmService
+from dataall.modules.dataset_sharing.services.dataset_sharing_alarm_service import DatasetSharingAlarmService
 from dataall.modules.dataset_sharing.db.share_object_models import ShareObjectItem, ShareObject
 from dataall.modules.dataset_sharing.services.share_managers.share_manager_utils import ShareErrorFormatter
 
@@ -584,7 +584,7 @@ class LFShareManager:
             f'due to: {error}'
         )
 
-        DatasetAlarmService().trigger_table_sharing_failure_alarm(table, self.share, self.target_environment)
+        DatasetSharingAlarmService().trigger_table_sharing_failure_alarm(table, self.share, self.target_environment)
         return True
 
     def handle_revoke_failure(
@@ -604,7 +604,9 @@ class LFShareManager:
             f'with target account {self.target_environment.AwsAccountId}/{self.target_environment.region} '
             f'due to: {error}'
         )
-        DatasetAlarmService().trigger_revoke_table_sharing_failure_alarm(table, self.share, self.target_environment)
+        DatasetSharingAlarmService().trigger_revoke_table_sharing_failure_alarm(
+            table, self.share, self.target_environment
+        )
         return True
 
     def handle_share_failure_for_all_tables(self, tables, error, share_item_status, reapply=False):
